Moved the methods of RemoteFilterDefinition from 'installations.go' to 'filter_remote.go'
